requirements.txt
django
django-allauth[socialaccount]
inngest 